discord-bot/requirements.txt
discord.py>=2.0.0
aiohttp
opuslib
webrtcvad
dotenv